requirements.txt
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0
pandas==2.2.2
openpyxl==3.1.4
